components/wr-motherduck/pyproject.toml
[project]
name = "wr-motherduck"
version = "0.0.8"
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.2.0",
    "freezegun>=1.5.1",
    "keboola-component>=1.6.10",
    "keboola-http-client>=1.0.1",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "pydantic>=2.11.3",
    "ruff>=0.11.5",
    "duckdb>=1.2.2",
    "kbcstorage>=0.9.2",
]
[tool.ruff]
line-length = 120